+++ src/syntax/SQLSugarSyntax.php
@@ -34,19 +34,27 @@ class SQLSugarSyntax
     return $this;
   }
 
+  public function getAllTheQueryList() {
+    return array_keys($this->osuql['queries']);
+  }
+
   public function getSQLObject() {
     $osuql = $this->osuql;
     $this->clear();
     return $osuql;
   }
 
-  public function getSQL() {
+  public function getSQL($queryList) {
+    if ($queryList === 'all')
+      $queryList = $this->getAllTheQueryList();
+
+    if (!is_array($queryList)) return null;
     if (!$this->adapter) return null;
 
     $classBuilder = SQLAdapter::get($this->adapter);
     $SQLBuilder = new $classBuilder($this->getSQLObject());
-    $SQLBuilder->run();
-    return $SQLBuilder->getSql();
+    $SQLBuilder->run($queryList);
+    return $SQLBuilder->getSql($queryList);
   }
 
   public function rel($leftTable, $rightTable, $on, $temporary = false) {