+++ tests/SQLSugarSyntaxTest.php
@@ -29,7 +29,7 @@ final class SQLSugarSyntaxTest extends TestCase
       'inner join user_group on users.id = user_group.user_id '.
       'inner join groups on user_group.group_id = groups.id '.
       'group by groups.name',
-      $db->getSQL()
+      $db->getSQL(['main'])
     );
 
     $this->assertEquals(